% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anomaly_index.R
\name{ai}
\alias{ai}
\title{Anomalie-Index}
\usage{
ai(x, ref = NULL)
}
\arguments{
\item{x}{ein data.frame mit Jahrringmessungen}

\item{ref}{ein data.frame mit Referenz-Jahrringmessungen (optional)}
}
\value{
ein data.frame
}
\description{
Anomalie-Index
}
\details{
Berechnet den Anomalie-Index. Details in Vorlesungs-Unterlagen.
}

